[DotNetCore] Refactor project template wizard
Move logic that determines the target frameworks from the page to
the wizard. This means the wizard can check the target frameworks
available and decide whether to show or hide the wizard page if
there are not enough frameworks available.
